# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tokyo-api backend project",
        "description": "backend for the tikshuv-logistics project",
        "contact": {
            "email": "your-email@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/clients": {
            "get": {
                "tags": [
                    "Clinet"
                ],
                "summary": "Get all clinets records name and id",
                "description": "Retrieves all clinets from the system only name and id.",
                "operationId": "d00d20521584ab5da83e088abb7aded2",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Percival Schulist"
                                            },
                                            "population": {
                                                "type": "string",
                                                "example": "c9810738"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "משתמש אינו מחובר"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "התרחש בעיית שרתת יש לנסות שוב מאוחר יותר."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/clients/search": {
            "get": {
                "tags": [
                    "Clinet"
                ],
                "summary": "Search for clients by personal number",
                "description": "Searches for clients in the system based on the provided personal number.",
                "operationId": "bb824283e4d62529812cc3b3079b6b02",
                "parameters": [
                    {
                        "name": "search_string",
                        "in": "path",
                        "description": "Search string (personal number)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Percival Schulist"
                                            },
                                            "personal_number": {
                                                "type": "string",
                                                "example": "s5671482"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "s5671482@army.idf.il"
                                            },
                                            "phone": {
                                                "type": "string",
                                                "example": "0556926412"
                                            },
                                            "emp_type_id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "employee_type": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "miluim"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "נתונים שנשנלחו אינם בפורמט תקין"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "BadRequest",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "חובה לשלוח מספר אישי לחיפוש"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "התרחשה בעיה בשרת. אנא נסה שוב מאוחר יותר."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}